Various updates

+++ update.sh
@@ -1,6 +1,13 @@
 #!/bin/bash
 
 cd /home/pw/bitcoin/stats
+
+if [ -f pidfile ]; then
+  exit
+fi
+
+echo $$ >pidfile
+
 rm *.png
 
 ./keepdump.pl >/dev/null
@@ -18,6 +25,7 @@ START=2016
 ./slide   1000                       $START <dump >slide-1000-ever.dat
 ./slide   500                        $START <dump >slide-500-ever.dat
 ./difflist.pl                        $START <dump >diff-ever.dat
+./verpct.pl                          $START <dump >ver-ever.dat
 
 START=$(($BLOCKS-50400))
 ./predict 2592000       $((3600*24)) $START <dump >predictM-50k.dat
@@ -29,6 +37,7 @@ START=$(($BLOCKS-50400))
 ./slide   500                        $START <dump >slide-500-50k.dat
 ./slide   200                        $START <dump >slide-200-50k.dat
 ./difflist.pl                        $START <dump >diff-50k.dat
+./verpct.pl                          $START <dump >ver-50k.dat
 
 START=$(($BLOCKS-10080))
 ./predict 2592000       $((3600*24)) $START <dump >predictM-10k.dat
@@ -40,6 +49,7 @@ START=$(($BLOCKS-10080))
 ./slide   500                        $START <dump >slide-500-10k.dat
 ./slide   200                        $START <dump >slide-200-10k.dat
 ./difflist.pl                        $START <dump >diff-10k.dat
+./verpct.pl                          $START <dump >ver-10k.dat
 
 START=$(($BLOCKS-2016))
 ./predict 2592000       $((3600*24)) $START <dump >predictM-2k.dat
@@ -51,11 +61,13 @@ START=$(($BLOCKS-2016))
 ./slide   500                        $START <dump >slide-500-2k.dat
 ./slide   200                        $START <dump >slide-200-2k.dat
 ./difflist.pl                        $START <dump >diff-2k.dat
+./verpct.pl                          $START <dump >ver-2k.dat
 
 START=$(($BLOCKS-1008))
 ./predict 86400         1200         $START 0.3  <dump >predictD-1k.dat
 ./predict 86400         1200         $START 0.06 <dump >predictDs-1k.dat
 ./difflist.pl                        $START <dump >diff-1k.dat
+./verpct.pl                          $START <dump >ver-1k.dat
 
 gnuplot growth-50k.gp
 gnuplot growth-10k.gp
@@ -76,7 +88,14 @@ gnuplot powdays-50k.gp
 gnuplot work-ever.gp
 gnuplot work-50k.gp
 
+gnuplot ver-2k.gp
+gnuplot ver-10k.gp
+gnuplot ver-50k.gp
+gnuplot ver-ever.gp
+
 tail -n 1 predict3D-10k.dat | cut -d ' ' -f 2 >speed-3D.txt
 
 cp *.png *.html *.css *txt /var/www/
 gzip --best <~/git/bitcoin-seeder/dnsseed.dump >/var/www/seeds.txt.gz
+
+rm -f pidfile
